Move meat of application into dedicated src folder
