+++ package.json
@@ -25,6 +25,7 @@
   "devDependencies": {
     "dotenv": "^16.1.4",
     "jest": "^29.6.1",
+    "mongodb-memory-server": "^8.13.0",
     "nodemon": "^2.0.22",
     "supertest": "^6.3.3"
   }
